dashboard/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
snowflake-connector-python>=3.6.0
